fix: fix invoice response

+++ src/main/java/com/swp/jewelrystore/model/dto/InvoiceDTO.java
@@ -26,5 +26,5 @@ public class InvoiceDTO {
     private String sellOrderStatus;
 
     @ApiModelProperty(example = "[92714, 89660]")
-    private List<Long> price;
+    private List<Double> price;
 }
